O.K

+++ src/main/java/com/ds/domain/SignupVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class SignupVO {
-	int lecture_no;
+	Long lecture_no;
 	String[] checkArr;
 	List<String> list;
 }
